final commit

+++ init.sh
@@ -14,23 +14,25 @@ chsh -s $(which zsh)
 # uv package manager
 curl -LsSf https://astral.sh/uv/install.sh | sh
 
-# LM studio
-mkdir exec
-wget https://installers.lmstudio.ai/linux/x64/0.3.20-4/LM-Studio-0.3.20-4-x64.AppImage -O exec/lms.AppImage
-chmod +x exec/lms.AppImage
-cd exec
-./lms.AppImage --appimage-extract
-mv squashfs-root/ lms/
-rm -rf lms.AppImage
-chmod +x lms/lm-studio
-nohup xvfb-run -a ./lms/lm-studio --no-sandbox > /dev/null 2>&1 < /dev/null &
-sleep 10
-echo -e "y\ny\ny" | npx --yes lmstudio install-cli
-export PATH="$PATH:/root/.lmstudio/bin"
-lms server start
-lms server stop
-sed -i 's/127.0.0.1/0.0.0.0/g' "/root/.lmstudio/.internal/http-server-config.json"
-sed -i 's/"enableLocalService": false/"enableLocalService": true/g' "/root/.config/LM Studio/settings.json"
+# LM studio - only run in headless mode
+if [ "$1" = "headless" ]; then
+    mkdir exec
+    wget https://installers.lmstudio.ai/linux/x64/0.3.20-4/LM-Studio-0.3.20-4-x64.AppImage -O exec/lms.AppImage
+    chmod +x exec/lms.AppImage
+    cd exec
+    ./lms.AppImage --appimage-extract
+    mv squashfs-root/ lms/
+    rm -rf lms.AppImage
+    chmod +x lms/lm-studio
+    nohup xvfb-run -a ./lms/lm-studio --no-sandbox > /dev/null 2>&1 < /dev/null &
+    sleep 10
+    echo -e "y\ny\ny" | npx --yes lmstudio install-cli
+    export PATH="$PATH:/root/.lmstudio/bin"
+    lms server start
+    lms server stop
+    sed -i 's/127.0.0.1/0.0.0.0/g' "/root/.lmstudio/.internal/http-server-config.json"
+    sed -i 's/"enableLocalService": false/"enableLocalService": true/g' "/root/.config/LM Studio/settings.json"
+fi
 
 # Final
 echo 'llm-se() { ~/fu27soma-ma/manage.sh "$@" ;}' >> ~/.zshrc
